final stage

+++ src/recipes/web/RecipeController.java
@@ -1,21 +1,18 @@
 package recipes.web;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.lang.Nullable;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-import recipes.domain.recipe.entity.Recipe;
-import recipes.domain.recipe.models.RecipeDto;
-import recipes.domain.recipe.models.ReturnId;
-import recipes.domain.recipe.service.RecipeService;
+import recipes.domain.entity.Recipe;
+import recipes.domain.models.RecipeDto;
+import recipes.domain.models.ReturnId;
+import recipes.domain.service.RecipeService;
 import javax.validation.Valid;
 import java.util.List;
-import java.util.Set;
 
 @RestController
 @RequestMapping("api/recipe")
@@ -25,8 +22,10 @@ public class RecipeController {
     private RecipeService recipeService;
 
     @GetMapping("{id}")
-    public ResponseEntity<RecipeDto> findById(@PathVariable Long id) {
+    public ResponseEntity<RecipeDto> getRecipe(@PathVariable Long id) {
+        log.info("get recipe before findbyid");
         RecipeDto recipe = recipeService.findById(id);
+        log.info("get recipe after find by id");
         return ResponseEntity.ok(recipe);
     }
 
@@ -36,27 +35,35 @@ public class RecipeController {
         return ResponseEntity.ok().body(createdRecipe);
     }
     @DeleteMapping("{id}")
-    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+    public ResponseEntity<?> deleteById(@PathVariable Long id, @AuthenticationPrincipal UserDetails userDetails) {
+        Recipe recipe = recipeService.findByIdRecipe(id);
+        if (!recipe.getUser().getEmail().equals(userDetails.getUsername())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         recipeService.deleteById(id);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
     @PutMapping("{id}")
-    public ResponseEntity<?> updateRecipe(@PathVariable Long id, @Valid @RequestBody RecipeDto recipe) {
+    public ResponseEntity<?> updateRecipe(@PathVariable Long id,
+                                          @Valid @RequestBody RecipeDto recipe,
+                                          @AuthenticationPrincipal UserDetails userDetails) {
+        Recipe recipeToUpdate = recipeService.findByIdRecipe(id);
+        if (!recipeToUpdate.getUser().getEmail().equals(userDetails.getUsername())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         recipeService.updateRecipe(id, recipe);
         log.info("update");
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @GetMapping
-    public ResponseEntity<Page<Recipe>> findAllRecipes(Pageable pageable) {
-       return ResponseEntity.ok(recipeService.findAllRecipes(pageable));
-    }
-
     @GetMapping("/search")
-    public List<RecipeDto> findAllByCategoryOrName(@RequestParam(required = false) String name, @RequestParam(required = false) String category) {
+    public List<RecipeDto> findAllByCategoryOrName(
+            @RequestParam(required = false, name = "name") String name,
+            @RequestParam(required = false,  name = "category") String category) {
         log.info("search");
         return recipeService.findAllByCategoryOrName(name, category);
 
     }
+
 
 }
